Refactored data classes
